Significant SVM improvements
File read/write during SVM scale/predict was causing the frame rate of
the app to drop considerably. This was originally done for debugging
the values from libSVM. The implementations have now been changed to
use NSMutableArrays rather than passing around references to files and
reading them in. The only files accessed now are the SVM model and the
corresponding scaling factors file.