iPhone settings all implemented
In addition to the settings implemented on the previous commit, the
classification settings have now been implemented for the iPhone app.
Now a user can choose which emotions to see predictions for, and the
camera view will respond accordingly. These settings should be
remembered when a user closes the app and opens again, but
classification will still need to be enabled.

A small amount of work is still remaining to update the iPad storyboard
and to link in the new settings code.